requirements.txt
streamlit
pandas
numpy
sentence-transformers
faiss-cpu
scikit-learn
openai
streamlit==1.38.0
